requirements.txt
requests
frida-tools